requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Data handling
pandas>=2.0.0
datasets>=2.14.0
pyarrow>=14.0.0

# Web scraping
httpx>=0.25.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
lxml>=4.9.0
html5lib>=1.1
readability-lxml>=0.8.1

# Text processing
nltk>=3.8.0
langdetect>=1.0.9
datasketch>=1.6.0
sentencepiece>=0.1.99

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Config & utilities
PyYAML>=6.0
tqdm>=4.66.0
click>=8.1.0
coloredlogs>=15.0

# Logging & monitoring
tensorboard>=2.14.0
wandb>=0.15.0

# Performance
accelerate>=0.24.0
